requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
configparser==5.0.0
dateutils==0.6.7
defusedxml==0.6.0
Django==3.0.2
django-cors-headers==3.2.1
django-rest-framework-braces==0.3.4
djangorestframework==3.11.0
facebook-sdk==3.1.0
gunicorn==20.0.4
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.4
pyjavaproperties==0.7
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rest-social-auth==3.0.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.8
